Fix corner case for switch
