+++ src/__tests__/loop.test.ts
@@ -265,60 +265,4 @@ describe('loop', () => {
     const expectedLogOutput = [{ binary: intToBinary(100), type: INT_BASE_TYPE }]
     expectLogOutputToBe(logOutput, expectedLogOutput)
   })
-
-  test('switch case without break', () => {
-    const output = testProgram(
-      `
-      int main() {
-        int x = 10;
-        int i = 0;
-        for (; i < x; i++)
-          switch(x)
-          {
-            case 1:
-              i = i + 2;
-            case 9:
-              i = i - 1;
-            default:
-              i = i + 1;
-          }
-        printfLog(i);
-        return 0;
-      }
-      `,
-    )
-    verifyProgramCompleted(output)
-    const logOutput = output.getLogOutput()
-    const expectedLogOutput = [{ binary: intToBinary(10), type: INT_BASE_TYPE }]
-    expectLogOutputToBe(logOutput, expectedLogOutput)
-  })
-
-  test('switch case without break', () => {
-    const output = testProgram(
-      `
-      int main() {
-        int x = 10;
-        int i = 0;
-        for (; i < x; i++)
-          switch(x)
-          {
-            case 1:
-              i = i + 2;
-              break;
-            case 9:
-              i = i - 1;
-              break;
-            default:
-              i = i + 1;
-          }
-        printfLog(i);
-        return 0;
-      }
-      `,
-    )
-    verifyProgramCompleted(output)
-    const logOutput = output.getLogOutput()
-    const expectedLogOutput = [{ binary: intToBinary(10), type: INT_BASE_TYPE }]
-    expectLogOutputToBe(logOutput, expectedLogOutput)
-  })
 })
